requirements-mcp.txt
# For running the MCP server (mcp_servers/max_server.py)
mcp>=1.0.0
anyio>=4.0.0
httpx>=0.27.0
duckduckgo_search>=6.0.0
pyyaml>=6.0
